feat(uart): add command line argument parser

+++ external/embarkos/package/embark-uart/include/uart.h
@@ -1,14 +1,17 @@
+#include <stddef.h>
+#include <sys/types.h>
 #ifndef UART_H
 #define UART_H
 
 #define UART_DEFAULT_DEVICE "/dev/ttyUSB0"
 #define UART_DEFAULT_BAUD   115200
 
+
 int uart_open(const char *device);
 
 int uart_configure(int fd, int baudrate);
 
-int uart_read(int fd, char *buffer, int size);
+ssize_t uart_read(int fd, char *buffer, size_t size);
 
 void uart_close(int fd);
